# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Lotus RFQ Service API
  version: 0.1.0
  description: |
    Static API contract for the Lotus RFQ service.

    This spec separates implemented callable APIs from planned or unregistered APIs.
    Do not call endpoints marked `x-lotus-callable: false`.

    Execution safety notes:
    - RFQ accept may trigger execution only if the lane is operator-approved.
    - Matcher-ready lanes are not executable authority.
    - Execution may fail closed.
    - Settlement verification may delay final completion.
    - Ghost-fill protection can change final state.
    - Accounting updates only after settlement/finality.
    - Execution-scope token is not authority by itself.

    Funding safety notes:
    - Funding v0 is non-custodial funding preparation: Lotus returns route instructions and tracks user-broadcast transactions; users sign with their own wallets.
    - Lotus does not custody, sign, broadcast user wallet transactions, pool funds, or internally allocate user balances in v0.
    - LI.FI route quote does not equal ready-to-trade.
    - Execution preflight requires READY_TO_TRADE.
    - Funding Capability Matrix decides target chain/token/destination for capital movement.
    - Market Capability Matrix remains separate and decides market lane routeability/operator approval.
    - Withdrawal v0 is also non-custodial: Lotus records withdrawal intents, route previews, and user-broadcast tx hashes only.
...
    - Predict.fun is not PredictIt.
servers:
...
paths:
...
  /auth/turnkey/exchange:
    post:
      tags: [Authentication]
      summary: Exchange a verified Turnkey session for a Lotus user JWT.
      description: Accepts Turnkey session identity fields and returns a Lotus bearer token for user-facing API calls. Raw Turnkey session tokens must not be logged or displayed.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: public-turnkey-session
      x-lotus-side-effects: issues-user-jwt
      x-lotus-danger: medium
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [turnkeySessionToken, turnkeyUserId, turnkeyOrganizationId]
              properties:
                turnkeySessionToken: { type: string }
                turnkeyUserId: { type: string }
                turnkeyOrganizationId: { type: string }
      responses:
        "200": { $ref: "#/components/responses/GenericObject" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Read persistent notifications for the authenticated user.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: none
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: cursor
          in: query
          required: false
          schema: { type: string, format: date-time }
      responses:
        "200": { $ref: "#/components/responses/GenericObject" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /notifications/{id}/read:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Notifications]
      summary: Mark one notification read for the authenticated user.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: updates-notification
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/GenericObject" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /notifications/read-all:
    post:
      tags: [Notifications]
      summary: Mark all notifications read for the authenticated user.
      x-lotus-status: implemented
      x-lotus-callable: true
      x-lotus-auth: user
      x-lotus-side-effects: updates-notifications
      x-lotus-danger: low
      security: [{ bearerAuth: [] }]
      responses:
        "200": { $ref: "#/components/responses/GenericObject" }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    GenericObject:
      description: Artifact/service-shaped response.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        code: { type: string }
        error: { type: string }
        reason: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...